refactor: move filter bottom sheet inline
- Removed separate filter file
- Simplified filter logic
- Kept state inside bottom sheet
